Move type referencer to visitor package
