+++ src/main/java/no/ssb/lds/graphql/schemas/visitors/GraphQLTypeReferencerVisitor.java
@@ -1,4 +1,4 @@
-package no.ssb.lds.graphql.schemas;
+package no.ssb.lds.graphql.schemas.visitors;
 
 import graphql.schema.GraphQLFieldDefinition;
 import graphql.schema.GraphQLInterfaceType;
